Create geometrical reinforcement
Here is a code to show that with a geometrical reinforcement, one way will be chosen at the detriment of the second one. The histograms are here to show that each way has a probability of 1/2 to be chosen. The last part is highliting the link between the reinforcement factor and the time n before one sequence becomes constant.